Update: better accuracy calculation
Former-commit-id: 79c94eb0099a8e88fe1fc5bf0ef9ba7138e5ba6c [formerly 7f245c09e27845c9323c952d09a04a144e5d4208]
Former-commit-id: c49c93c885065f259af490bcccc35c0b86d2f53d

+++ Quaver/src/Gameplay/GameplayRendering/GameplayUI.cs
@@ -199,7 +199,7 @@ namespace Quaver.Gameplay.GameplayRendering
             ScoreText.Text = Util.ScoreToString((int)CurrentScore);
 
             // Update Accuracy Text
-            CurrentAccuracy = Util.Tween(ScoreManager.Accuracy, (float)CurrentAccuracy, dt/10);
+            CurrentAccuracy = Util.Tween((float)ScoreManager.Accuracy, (float)CurrentAccuracy, dt/10);
             AccuracyCountText[0].Text = $"{CurrentAccuracy * 100:0.00}%";
 
             // Update Boundary
